+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-10 22:59:41
-// Commit: docs(state): add detailed code comments and architecture guide for state
+// Updated at: 2026-05-12 12:11:57
+// Commit: feat(state): integrate offline caching for state state
 
 class stateComponentState {
-  final String commitRef = '1044';
-  final String lastMessage = 'docs(state): add detailed code comments and architecture guide for state';
-  final DateTime timestamp = DateTime.parse('2026-05-10T22:59:41+05:00');
+  final String commitRef = '1061';
+  final String lastMessage = 'feat(state): integrate offline caching for state state';
+  final DateTime timestamp = DateTime.parse('2026-05-12T12:11:57+05:00');
 }